requirements.txt
requests
chromadb
sentence-transformers
psycopg2-binary
numpy
faiss-cpu
python-dotenv
openai
